tools/requirements.txt
pyserial
pytz
tzlocal